OT-2886 CalcTimer moved to separate file
